+++ Tests/Unit/Model/FrontEndUserGroupTest.php
@@ -4,6 +4,7 @@ declare(strict_types=1);
 
 namespace OliverKlee\Oelib\Tests\Unit\Model;
 
+use OliverKlee\Oelib\Model\AbstractModel;
 use OliverKlee\Oelib\Model\FrontEndUserGroup;
 use TYPO3\TestingFramework\Core\Unit\UnitTestCase;
 
@@ -24,63 +25,11 @@ final class FrontEndUserGroupTest extends UnitTestCase
         $this->subject = new FrontEndUserGroup();
     }
 
-    ////////////////////////////////
-    // Tests concerning getTitle()
-    ////////////////////////////////
-
     /**
      * @test
      */
-    public function getTitleForNonEmptyGroupTitleReturnsGroupTitle(): void
+    public function isModel(): void
     {
-        $this->subject->setData(['title' => 'foo']);
-
-        self::assertSame(
-            'foo',
-            $this->subject->getTitle()
-        );
-    }
-
-    /**
-     * @test
-     */
-    public function getTitleForEmptyGroupTitleReturnsEmptyString(): void
-    {
-        $this->subject->setData(['title' => '']);
-
-        self::assertSame(
-            '',
-            $this->subject->getTitle()
-        );
-    }
-
-    //////////////////////////////////////
-    // Tests concerning getDescription()
-    //////////////////////////////////////
-
-    /**
-     * @test
-     */
-    public function getDescriptionForNonEmptyGroupDescriptionReturnsGroupDescription(): void
-    {
-        $this->subject->setData(['description' => 'foo']);
-
-        self::assertSame(
-            'foo',
-            $this->subject->getDescription()
-        );
-    }
-
-    /**
-     * @test
-     */
-    public function getDescriptionForEmptyGroupDescriptionReturnsEmptyString(): void
-    {
-        $this->subject->setData(['description' => '']);
-
-        self::assertSame(
-            '',
-            $this->subject->getDescription()
-        );
+        self::assertInstanceOf(AbstractModel::class, $this->subject);
     }
 }